requirements.txt
Faker
psycopg2
mysql-connector-python